Correct IMU behavior following REP (ROS Enhancement Proposal) 103
The accelerometer, gyroscope, and magnetometer required a reconfiguration of values in order to match the coordinate conventions expected by ROS via REP 103. In particular, the x and y value must be swapped, and the x value must be negated. Once these changes are made, all measurements now follow the right hand rule and increase as the robot rotates counter-clockwise, and yaw is zero when pointing east. Additionally, accelerometer values are now fed to ROS as meters per second squared, gyroscope values as radians per second, and magnetometer values as radians.