Calculate linear and angular velocity; Correct odometry axis orientation
The Odometry class now tracks clock time and updates x, y, and theta velocity variables vx, vy, and vtheta. The odometry calculations have also been corrected to match REP 103, such that x increases as the robot moves forward, and y and theta increase as the robot moves or rotates to the left.