Bug fix: main sketch pin swap required encoder interrupt restructuring
Right encoder B pin is now at pin 8, which requires a Pin Change interrupt (swapped with left encoder B pin). Note that this is an extremely brittle factoring that should be restructured for increased flexibility; restructuring will most likely prove difficult because of the inherently deficient nature of Pin Change interrupts on the Arduino platform.