Enhance encoder behavior to improve localization
Calculating rotation from wheel encoders on a four-wheeled skid-steer robot seems to be highly error-prone. As a result, the odometry calculation now uses yaw value from IMU instead of estimating rotation via encoders. Additionally, odometry is not reported when rotating (instead, zeros are passed to ROS). This should be dealt with in a less kludgy way in the future.