Implement watchdog timer to mitigate failure of communication link
The Arduino now keeps track of when the last serial communication was received from the Intel NUC. If too much time has passed since the last message was received, the motors are automatically stopped to avoid the robot running loose in the event of a lengthy communication delay or termination of the link. This "watchdog timer" is currently set to 1 second, but can easily be changed by modifying the value in the preamble.