Implement two-servo operation to facilitate gripper control via serial command (i.e. ROS)
Serial command parser now accepts commands to open/close the gripper fingers ("f") on pin 9, and to raise/lower the gripper wrist ("w") on pin 12. Note that proper servo control requires modification of the Arduino Servo library to force Servo to use Timer3 instead of Timer1 (to avoid conflicts with the left-side motor PWM on pin 10).