Add customized version of Servo library to accommodate gripper pins
The Arduino IDE's built-in Servo library utilizes Timer1 by default to control any attached servo motors. Timer1 is hard-wired to communicate servo control signals (PWM) using pins 9, 10, and 11, which creates a conflict with the motor control signals that are also on pins 10 and 11. The Servo library has been modified in order to mitigate this conflict by switching from Timer1 to Timer3, which uses pin 5 (one of two pins connected to the gripper servos). To ensure that this modification persists over time, a bare bones copy of the built-in Servo code has been included as a user-defined library (along with the Timer1->Timer3 swap), which overrides the built-in library.