Add IMU diagnostic function to test I2C connectivity
This function is a stripped-down version of the i2c_scanner example sketch (http://playground.arduino.cc/Main/I2cScanner). In the future, the function could be enhanced to check for connectivity on specific I2C addresses, rather than assuming that a response on any address indicates a valid IMU connection.