Enhanced gripper servo controls
Added 4 variables to set the max and min PWM signal for the wrist and finger servos.
Replaced servo.write with servo.writeMicroseconds to improve consitancy and ease of moddifing the servo min, max travels.
There is now a line that converts from degrees to PWM at a liner ratio based upon the min and max settings for that servo.